server/ca_assist/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
pydantic[email]
langchain
langchain-community
langchain-core
langchain-ollama
langchain-huggingface
langchain-chroma
chromadb
sentence-transformers
PyPDF2
pytesseract
Pillow
openai
httpx
pytest
torch>=2.0.0
torchvision>=0.17.0
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
PyJWT
